set the default ENABLE_SWIG to OFF and not even give the option if CSHARP, PYTHON2, or OCTAVE interfaces are enabled.  move the definitions of the options to the interfaces folder CMAKE and create a DISABLE_C_SHARED_LIB option to explicily disable the C_SHARED_LIB otherwise by default it is enabled.

+++ src/helics/core/HandleManager.cpp
@@ -93,7 +93,7 @@ void HandleManager::addHandleAtIndex (const BasicHandleInfo &otherHandle, int32_
     }
     else if (index > 0)
     {
-        handles.resize (index + 1);
+        handles.resize (static_cast<size_t>(index) + 1);
         // use placement new to reconstruct new object
         new (&handles[index]) BasicHandleInfo (otherHandle);
         addSearchFields (handles[index], index);
